Don't crash marker on hosts without OVS
OVS marker should be able to live on hosts with OVS installed without
getting caught in a restart loop due to missing OVS socket.

With this change, we check if OVS is installed on the host. If it is not,
we just sleep forever. The functionality works fine, if no OVS bridge is
available, we just don't report any bridges.

Also be less graceful about failures during Update() to improve self-healing.
If our Kubernetes client is not valid anymore, it is better to restart the
marker and refresh the client.

Signed-off-by: Petr Horacek <phoracek@redhat.com>