ovs-test, Fix pods exists test
Currently It(assert pods exists) test checks if there is any pod,
to check if ovs-cni pod is up and running.
Since there is already a default pod named local-volume-provisioner in the cluster,
we need to make this test more specific to catch ovs-cni pod.
we also change from It() to Specify for better readability

Signed-off-by: Ram Lavi <ralavi@redhat.com>