[Release-0.19] ovs-cni, Move images to quay.io (#185)
* ovs-cni, Bump kubevirtci

As part of the effort to point to quay.io
instead of docker.io, in order to overcome the
rate limits,
bump kubevirtci k8s-1.17, because its images
are already on quay.io.

Signed-off-by: Or Shoval <oshoval@redhat.com>

* ovs-cni, Move images to quay.io

In order to fix the rate limit of docker.io,
update the images which are not part of cluster-up
to be on quay.io instead on docker.io.

Signed-off-by: Or Shoval <oshoval@redhat.com>

* ovs-cni, Freeze multus yaml

In order to freeze a specific multus yaml,
and also point its image to quay.io,
freeze multus yaml, and point to
the quay.io which is maintained by CNAO.

The vanilla yaml was fetched from
https://raw.githubusercontent.com/intel/multus-cni/master/images/multus-daemonset.yml

Signed-off-by: Or Shoval <oshoval@redhat.com>