main: increase update timeout loop
In order to reduce the ovs-bridge-marker net consumption, This commit reduces the number
of API calls by increasing the default poll-interval.

The marker test is updated and it's timeout is increased accordingly,
with same pattern as it was before this commit- two times more than the update
interval which is 240 seconds.

Signed-off-by: alonsadan <asadan@redhat.com>