Reduce net consumption (#186)
* main: increase update timeout loop

In order to reduce the ovs-bridge-marker net consumption, This commit reduces the number
of API calls by increasing the default poll-interval.

The marker test is updated and it's timeout is increased accordingly,
with same pattern as it was before this commit- two times more than the update
interval which is 240 seconds.

Signed-off-by: alonsadan <asadan@redhat.com>

* marker, reported-resources: use cache

In order to reduce the ovs-bridge-marker net consumption, this commit reduces the
number of API calls by adding a cache to keep track of the bridges already been
reported, and update the node object only every 10 minutes.

Signed-off-by: alonsadan <asadan@redhat.com>

* main: add jitter to update & reconcile

To avoid communication peeks, this commit adds a jitter to randomize
the duration it takes to update and retrieve info from the node.

Signed-off-by: alonsadan <asadan@redhat.com>