tests, marker: increase timeout (#189)
The current timeout in the tests is 2 minutes. The update interval is 1 minute,
and it has a jitter of 1.2.
IF the test assertion start right after the start of the interval, 1 minute has passed ,
and for some reason the update fails, there could be more than two minutes between updates,
and the tests could fail.

This commit increases the tests timeout to 3 minutes, which should be enough.

Signed-off-by: alonsadan <asadan@redhat.com>