Upgrade multus to the latest release
We currently use an ancient version of Multus.

With this patch, we will follow the latest stable release. As of today
that is 3.9.

Assuming Multus keeps their released images stable, this patch removes
the cached image in favor of the official channel.

Note that this also switches from thin Multus CNI to the thick version
of the plugin.

Signed-off-by: Petr Horáček <phoracek@redhat.com>