Wait for tcpdump to start listening
Installation and execution of tcpdump in the mirroring test is executed
async. On a slower setup, the installation of tcpdump may take more time
than the execution of ping inside the test - causing missed ping reply
and failing the test.

With this patch, we wait for tcpdump to start recording traffic before
the test continues its execution.

This pathc also removes waiting for the tcpdump file to get created,
assuming it is not needed with the wait now happening in preparation
phase.

Signed-off-by: Petr Horáček <phoracek@redhat.com>