Use two-stage container build (#247)
* Use two-stage container build

In the current code, binaries are built locally to be only then put into
the container image.

This patch aims to simplify part of that process, by containerizing
builds. This should also help us avoid environment related issues, and
later allow cross-platform building.

Signed-off-by: Petr Horáček <phoracek@redhat.com>

* Call the correct version script

The current code was using a library version.sh instead of executable
get_version.sh to render a file with version name.

This patch corrects it and it converts the library to a non-executable,
to prevent other mistakes similar to this.

Signed-off-by: Petr Horáček <phoracek@redhat.com>

Signed-off-by: Petr Horáček <phoracek@redhat.com>