Cleanup ports after IPAM failure
While most of the configuration gets cleaned up on ADD failure through
the removal of containers netns, OVS ports do not.

In case port attachment succeeded but the following IPAM configuration
failed, we end up with the port left behind.

With this patch, the port gets explicitly deleted on IPAM failure.

Signed-off-by: Petr Horáček <phoracek@redhat.com>