Add SF support
This patch adds support for scalable functions
(https://docs.nvidia.com/networking/display/bluefielddpuosv385/scalable+functions).
They are represented on the host as auxiliary devices.

In order to for the minimal codebase changes, this patch updated
the sriovnet dependency, which already has support for SFs, together
with its needed dependencies.

This patch also adds architecture detection to the Makefile.

Signed-off-by: Alin Gabriel Serdean <aserdean@nvidia.com>