GetOvsPortForContIface should not return error if iface not found
Currently the function returns '"",false, err' in case if interface
not found.

The signature and usage of this function assume that the
function should not return and error in case if interface not
found.

Fix the function to return '"", false, nil' in case if the interface
not found.

This change fixes CmdDel for scenario when CmdDel is called
after failed CmdAdd or when port was manually removed from the ovs.

Signed-off-by: Yury Kulazhenkov <ykulazhenkov@nvidia.com>