Add the Pod's UID to the external_ids field of OVSDB
**What this PR does / why we need it**:

By adding the UID of a Pod to the external_ids field of the
corresponding veth interface in the OVSDB, it becomes convenient for the
Pod object to query the OVSDB using the UID to find the vethName of the
Pod on the host machine, and then deliver OVS flow tables accordingly.

Currently, the host machine's veth interface in the OVSDB only stores
the Netns field, which is useful, but for Pod objects, their netns is
not publicly accessible. This change addresses this limitation by
allowing Pod objects to query the OVSDB directly using their UID.

**Special notes for your reviewer**:

**Release note**:

```release-note
Added the Pod UID to the external_ids field of the corresponding veth
interface in the OVSDB.
This change allows querying the vethName of a Pod on the host machine
using its UID, improving the ability to deliver OVS flow tables for Pod
objects.
```

Signed-off-by: jiayoukun <824807548@qq.com>